fix(plugin): deduplicate PluginTier + move weightToTier imports to tier.ts
#722 squash + #724 left duplicate PluginTier type declarations and
weightToTier in both types.ts and tier.ts. Clean up:
- Remove duplicate PluginTier + weightToTier from types.ts
- Update plugin-list-manifest.ts and plugins-ls-info.ts to import
  weightToTier from tier.ts (avoids bun CI resolution bug)
